Finished test ant target + temporarly removed build directory executions

+++ install.sh
@@ -11,6 +11,11 @@ if [ ! -f $buildFile ]; then
     echo "buildFile $buildFile does not exist"
 fi
 
+if [ -f $linkName ]; then
+    echo "Link already exists"
+    exit 0
+fi
+
 ln -s $buildFile $linkName
 if [ -f $linkName ]; then
     echo "Link was succesfuly created"
